Backend/backApp/src/main/java/com/backend/backApp/service/CreateAccountService.java
package com.backend.backApp.service;

import com.backend.backApp.models.Account;
import com.backend.backApp.models.AccountDTO;
import com.backend.backApp.repository.AccountRepository;
import com.backend.backApp.validators.AccountValidator;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;


@Service
public class CreateAccountService implements Command<Account, AccountDTO>{

    private final AccountRepository accountRepository;

    public CreateAccountService(AccountRepository accountRepository) {
        this.accountRepository = accountRepository;
    }

    @Override
    public ResponseEntity<AccountDTO> execute(Account account) {
        AccountValidator.execute(account);
        Account savedAccount = accountRepository.save(account);
        return ResponseEntity.status(HttpStatus.CREATED).body(new AccountDTO(savedAccount));
    }
}
